+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - PASSBOLT-2667: Fixed regression: create a user that has been deleted previously returns an error
 - PASSBOLT-2673: Fixed regression: as AD I cannot create a group with the name of previously deleted group
 - PASSBOLT-2545: Fixed regression: As AD deleting a group I should be notified that all members of the group gonna lose access to the passwords shared with the group
+- PASSBOLT-2597: Before migrating to 2.0 ensure passbolt is already in latest 1.x
 
 ### Added
 - PASSBOLT-2638: Added command to test email configuration and SMTP communication